refactor(Components): Standardize components name with Capital Letter
