+++ src/components/PaletteList.js
